requirements.txt
# Python dependencies for mxi-samsung ETL project
# ============================================================================

# ============================================================================
# Web Scraping & API
# ============================================================================
requests==2.31.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# ============================================================================
# Data Processing & Analysis
# ============================================================================
pandas==2.1.1
numpy==1.24.3
pyarrow==13.0.0  # For Parquet support

# ============================================================================
# Database & ORM
# ============================================================================
psycopg2-binary==2.9.9  # PostgreSQL adapter
sqlalchemy==2.0.23

# ============================================================================
# Configuration & Environment
# ============================================================================
python-dotenv==1.0.0
pydantic==2.5.0

# ============================================================================
# Development & Testing
# ============================================================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
pylint==3.0.2

# ============================================================================
# Airflow core (optional - included in Docker image)
# ============================================================================
# apache-airflow==2.7.0
# apache-airflow-providers-postgres==5.10.0
# apache-airflow-providers-http==4.7.0
